Refactor statement components into individual files for better modularity
Split the large components.tsx file into 18 individual statement component files,
following the same pattern used for expressions. This improves code maintainability
and follows the established modular architecture.

Changes:
- Create individual files for each statement component type
- Update main components.tsx to import from individual files
- Fix TypeScript imports and remove circular dependency issues  
- Add React import to types.ts for proper React.FC typing
